+++ kamehouse-shell/bin/kamehouse/get-video-playlist-content.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# Disable logs
+LOG=ERROR
+SKIP_LOG_START_FINISH=true
+
 # Import kamehouse functions
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then
@@ -7,7 +11,6 @@ if [ "$?" != "0" ]; then
   exit 99
 fi
 
-SKIP_LOG_START_FINISH=true
 PLAYLIST_FILE=""
 
 mainProcess() {